requirements.txt
fastapi
uvicorn
python-dotenv
google-auth
google-auth-oauthlib
google-api-python-client
pydantic
